HBASE-29223 Move MasterStatusConstants and MasterStatusUtil
from org.apache.hadoop.hbase.util to org.apache.hadoop.hbase.master.http package. Reason: these are exclusively used by the web UI.
