+++ hbase-server/src/main/java/org/apache/hadoop/hbase/master/http/MasterStatusConstants.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.hadoop.hbase.util;
+package org.apache.hadoop.hbase.master.http;
 
 import org.apache.yetus.audience.InterfaceAudience;
