+++ hbase-server/src/main/resources/hbase-webapps/master/regionServerListBaseStats.jsp
@@ -27,8 +27,8 @@
          import="java.util.*"
          import="org.apache.hadoop.hbase.ServerMetrics"
          import="org.apache.hadoop.util.StringUtils"
-         import="org.apache.hadoop.hbase.util.MasterStatusConstants"
-         import="org.apache.hadoop.hbase.util.MasterStatusUtil" %>
+         import="org.apache.hadoop.hbase.master.http.MasterStatusConstants"
+         import="org.apache.hadoop.hbase.master.http.MasterStatusUtil" %>
 <%
   ServerName[] serverNames = (ServerName[]) request.getAttribute(MasterStatusConstants.SERVER_NAMES);
   HMaster master = (HMaster) getServletContext().getAttribute(HMaster.MASTER);
