+++ hbase-server/src/main/resources/hbase-webapps/master/regionServerListMemoryStats.jsp
@@ -24,8 +24,8 @@
          import="org.apache.hadoop.util.StringUtils.TraditionalBinaryPrefix"
          import="org.apache.hadoop.hbase.RegionMetrics"
          import="org.apache.hadoop.hbase.Size"
-         import="org.apache.hadoop.hbase.util.MasterStatusConstants"
-         import="org.apache.hadoop.hbase.util.MasterStatusUtil" %>
+         import="org.apache.hadoop.hbase.master.http.MasterStatusConstants"
+         import="org.apache.hadoop.hbase.master.http.MasterStatusUtil" %>
 
 <%
   ServerName[] serverNames = (ServerName[]) request.getAttribute(MasterStatusConstants.SERVER_NAMES);
