+++ hbase-server/src/main/resources/hbase-webapps/master/rsGroupList.jsp
@@ -27,7 +27,7 @@
          import="org.apache.hadoop.hbase.ServerMetrics"
          import="org.apache.hadoop.hbase.net.Address"
          import="org.apache.hadoop.hbase.rsgroup.RSGroupInfo"
-         import="org.apache.hadoop.hbase.util.MasterStatusConstants" %>
+         import="org.apache.hadoop.hbase.master.http.MasterStatusConstants" %>
 
 <%
   HMaster master = (HMaster) getServletContext().getAttribute(HMaster.MASTER);
