+++ hbase-server/src/main/resources/hbase-webapps/master/rsGroupListStoreStats.jsp
@@ -25,7 +25,7 @@
          import="org.apache.hadoop.util.StringUtils.TraditionalBinaryPrefix"
          import="org.apache.hadoop.hbase.RegionMetrics"
          import="org.apache.hadoop.hbase.Size"
-         import="org.apache.hadoop.hbase.util.MasterStatusConstants" %>
+         import="org.apache.hadoop.hbase.master.http.MasterStatusConstants" %>
 
 <%
   RSGroupInfo [] rsGroupInfos = (RSGroupInfo[]) request.getAttribute(MasterStatusConstants.RS_GROUP_INFOS);
